Add cop DefEndAlignment
Move checks of def-end from EndAlignment into a new cop.
