+++ lib/rubocop.rb
@@ -55,6 +55,7 @@ require 'rubocop/cop/mixin/configurable_max'
 require 'rubocop/cop/mixin/code_length'
 require 'rubocop/cop/mixin/configurable_enforced_style'
 require 'rubocop/cop/mixin/configurable_naming'
+require 'rubocop/cop/mixin/end_keyword_alignment'
 require 'rubocop/cop/mixin/if_node'
 require 'rubocop/cop/mixin/if_then_else'
 require 'rubocop/cop/mixin/negative_conditional'
@@ -74,6 +75,7 @@ require 'rubocop/cop/lint/assignment_in_condition'
 require 'rubocop/cop/lint/block_alignment'
 require 'rubocop/cop/lint/condition_position'
 require 'rubocop/cop/lint/debugger'
+require 'rubocop/cop/lint/def_end_alignment'
 require 'rubocop/cop/lint/deprecated_class_methods'
 require 'rubocop/cop/lint/else_layout'
 require 'rubocop/cop/lint/empty_ensure'
